animated particle effects

+++ graphics/src/main/resources/pl/warp/engine/graphics/shader/particle/vert.glsl
@@ -7,7 +7,7 @@ uniform mat4 modelViewMatrix;
 
 layout(location = 0) in vec3 position;
 layout(location = 1) in float rotation;
-layout(location = 2) in int textureIndex;
+layout(location = 2) in float textureIndex;
 
 out vData {
     mat2 rotation;
